+++ src/main/java/com/jazeee/ddp/client/DdpSubscriptionNotificationProcessor.java
@@ -1,4 +1,4 @@
-package com.keysolutions.ddpclient;
+package com.jazeee.ddp.client;
 
 import com.jazeee.common.notifier.INotificationProcessor;
 import com.jazeee.common.utils.nullability.NotNull;
