+++ src/test/java/com/jazeee/ddp/client/AllTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.keysolutions.ddpclient.test;
+package com.jazeee.ddp.client;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
