+++ src/test/java/com/jazeee/ddp/client/TestDDPConnections.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.keysolutions.ddpclient.test;
+package com.jazeee.ddp.client;
 
 import java.lang.reflect.Method;
 import java.net.URISyntaxException;
 
 import junit.framework.TestCase;
 
-import com.keysolutions.ddpclient.DdpClient;
-import com.keysolutions.ddpclient.test.DdpTestClientListener.DdpState;
+import com.jazeee.ddp.client.DdpClient;
+import com.jazeee.ddp.client.DdpTestClientListener.DdpState;
 
 public class TestDDPConnections extends TestCase {
